PMTOK: Reorg files and classes
